Add plot interpretation to write up
This is so we can give a clear idea on how this work can be applied to the business. Moreover, add intuition for function to get the closest pair of factors to a number so it is easier for people to understand the logic.

Move the text cleaning function out of the wrangling code and into its own script so test can be done on it.